+++ docker-compose.yml
@@ -4,7 +4,7 @@ services:
     image: python:3.6.3-onbuild
     command: bash
     volumes:
-      - .:/usr/src/app
+      - .:/usr/longitude/app
 
   cache:
     image: redis:alpine
