+++ longitude/core/caches/base.py
@@ -2,7 +2,7 @@ import hashlib
 import logging
 import pickle
 
-from src.core.common.query_response import LongitudeQueryResponse
+from longitude.core.common.query_response import LongitudeQueryResponse
 from ..common.config import LongitudeConfigurable
