+++ longitude/core/tests/test_cache_redis.py
@@ -2,11 +2,11 @@ from unittest import TestCase, mock
 
 import redis.exceptions
 
-from src.core.common.query_response import LongitudeQueryResponse
+from longitude.core.common.query_response import LongitudeQueryResponse
 from ..caches.redis import RedisCache
 
 
-@mock.patch('src.core.caches.redis.redis.Redis')
+@mock.patch('longitude.core.caches.redis.redis.Redis')
 class TestRedisCache(TestCase):
     cache = None
 
@@ -39,7 +39,7 @@ class TestRedisCache(TestCase):
         with self.assertLogs(level='ERROR') as log_test:
             self.assertFalse(self.cache.is_ready)
             expected_log = [
-                'ERROR:src.core.caches.redis:Cannot connect to Redis server at some_host:666.'
+                'ERROR:longitude.core.caches.redis:Cannot connect to Redis server at some_host:666.'
             ]
 
             self.assertEqual(expected_log, log_test.output)
@@ -54,18 +54,18 @@ class TestRedisCache(TestCase):
         self.cache.setup()
         with self.assertLogs(level='ERROR') as log_test:
             self.assertFalse(self.cache.is_ready)
-            self.assertEqual(['ERROR:src.core.caches.redis:Redis password required.'], log_test.output)
+            self.assertEqual(['ERROR:longitude.core.caches.redis:Redis password required.'], log_test.output)
 
     def test_is_not_ready_because_wrong_password(self, redis_mock):
         redis_mock.return_value.ping.side_effect = redis.exceptions.ResponseError('invalid password')
         self.cache.setup()
         with self.assertLogs(level='ERROR') as log_test:
             self.assertFalse(self.cache.is_ready)
-            self.assertEqual(['ERROR:src.core.caches.redis:Redis password is wrong.'], log_test.output)
+            self.assertEqual(['ERROR:longitude.core.caches.redis:Redis password is wrong.'], log_test.output)
 
     def test_is_not_ready_because_of_generic_response_error(self, redis_mock):
         redis_mock.return_value.ping.side_effect = redis.exceptions.ResponseError('some error text')
         self.cache.setup()
         with self.assertLogs(level='ERROR') as log_test:
             self.assertFalse(self.cache.is_ready)
-            self.assertEqual(['ERROR:src.core.caches.redis:some error text'], log_test.output)
+            self.assertEqual(['ERROR:longitude.core.caches.redis:some error text'], log_test.output)
