+++ longitude/core/tests/test_helpers.py
@@ -1,11 +1,11 @@
 from unittest import TestCase, mock
-from src.core.common.helpers import DisabledCache
-from src.core.data_sources.base import DataSource
+from longitude.core.common.helpers import DisabledCache
+from longitude.core.data_sources.base import DataSource
 
 
 class TestHelpers(TestCase):
 
-    @mock.patch('src.core.data_sources.base.DataSource', spec=DataSource)
+    @mock.patch('longitude.core.data_sources.base.DataSource', spec=DataSource)
     def test_disable_cache_context_manager_triggers_cache(self, fake_data_source):
         fake_data_source.enable_cache.return_value = None
         fake_data_source.disable_cache.return_value = None
@@ -13,7 +13,7 @@ class TestHelpers(TestCase):
             fake_data_source.disable_cache.assert_called_once()
         fake_data_source.enable_cache.assert_called_once()
 
-    @mock.patch('src.core.data_sources.base.DataSource')
+    @mock.patch('longitude.core.data_sources.base.DataSource')
     def test_disable_cache_context_manager_must_receive_a_data_source(self, fake_data_source):
         with self.assertRaises(TypeError):
             with DisabledCache(fake_data_source):
