+++ longitude/samples/carto_sample_with_ram_cache.py
@@ -10,13 +10,16 @@ You must create a 'carto_sample_config.py' file at this folder with the needed f
 That file will be ignored in git, so do not worry about pushing credentials anywhere (but BE CAREFUL!)
 DO NOT REPLACE THIS WITH HARD CODED CREDENTIALS EVER AND ALWAYS REVIEW YOUR COMMITS!
 """
+
+import time
 import os
 import sys
 
 sys.path.append(os.path.join(os.path.dirname(__file__), '..', '..'))
-from src.core.data_sources.base import LongitudeRetriesExceeded
-from src.core.data_sources.carto import CartoDataSource
-from src.samples.carto_sample_config import CARTO_API_KEY, CARTO_USER, CARTO_TABLE_NAME
+from longitude.core.caches.ram import RamCache
+from longitude.core.data_sources.base import LongitudeRetriesExceeded
+from longitude.core.data_sources.carto import CartoDataSource
+from longitude.samples.carto_sample_config import CARTO_API_KEY, CARTO_USER, CARTO_TABLE_NAME
 
 if __name__ == "__main__":
     config = {
@@ -24,13 +27,47 @@ if __name__ == "__main__":
         'user': CARTO_USER
     }
 
-    ds = CartoDataSource(config)
+    ds = CartoDataSource(config, cache_class=RamCache)
     ds.setup()
     if ds.is_ready:
         try:
-            data = ds.query('select * from %s limit 30' % CARTO_TABLE_NAME)
-            [print(r) for r in data.rows]
-            print(data.profiling)
+
+            REPEATED_QUERY = 'select * from %s limit 30' % CARTO_TABLE_NAME
+            start = time.time()
+            data = ds.query(REPEATED_QUERY)
+            elapsed = time.time() - start
+            print("It took %s without cache" % elapsed)
+            print('Uses cache? ' + str(data.comes_from_cache))
+
+            # Repeated read queries return cached values
+            start_with_cache = time.time()
+            cached_data = ds.query(REPEATED_QUERY)
+            elapsed_with_cache = time.time() - start_with_cache
+            print("It took %s with cache" % elapsed_with_cache)
+            print('Uses cache? ' + str(cached_data.comes_from_cache))
+
+            # Data is the same...
+            assert str(data) == str(cached_data)
+
+            # You can also disable the cache for a while (nothing gets read or written)
+            ds.disable_cache()
+            start = time.time()
+            data = ds.query(REPEATED_QUERY)
+            elapsed = time.time() - start
+            print('It took %s with disabled cache' % str(elapsed))
+            print('Uses cache? ' + str(data.comes_from_cache))
+            ds.enable_cache()
+
+            # Or disable specific queries via query_config (nothing gets read or written)
+            query_config = ds.copy_default_query_config()
+            start = time.time()
+            data = ds.query(REPEATED_QUERY, query_config=query_config, use_cache=False)
+            elapsed = time.time() - start
+            print('It took %s with disabled cache (per-query)' % str(elapsed))
+            print('Uses cache? ' + str(data.comes_from_cache))
+
+            print('If you see decreasing times it is probably because CARTOs cache doing its job!')
+
         except LongitudeRetriesExceeded:
             print("Too many retries and no success...")
     else:
